refactor dir structure and clean up code
- move landing page features to src/features
- remove unnecessary comments
- add more meaningful text placeholders
- appease the linter
